Refactor options & tags/options as TypeScript
